LangChain-Projects/rag_chatbot/pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "abdulrehman1004", email = "rehman.abdul008844@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.17",
    "langchain-community>=0.3.16",
    "langchain-core>=0.3.33",
    "langchain-google-genai>=2.0.9",
    "langchain-pinecone>=0.2.2",
    "pinecone-notebooks>=0.1.1",
    "pypdf>=5.2.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.41.1",
]

[project.scripts]
rag-chatbot = "rag_chatbot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
